projects/172-thr-disclosure-parse/requirements.txt
# THR Disclosure Parse Project Requirements
# DA-172: Extract and Join COI Data

# Google Cloud Libraries
google-cloud-bigquery>=3.11.0
google-cloud-firestore>=2.11.0
google-auth>=2.16.0

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Excel export support
openpyxl>=3.1.0  # Excel file reading/writing
xlsxwriter>=3.1.0  # Advanced Excel formatting and features